requirements_backend.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.12.3
starlette==0.48.0
sqlalchemy==2.0.44
aiomysql==0.3.2
alembic==1.17.0
cryptography==46.0.3
PyMySQL==1.1.2


torch==2.9.0
torchvision==0.24.0
pytorch-lightning==2.5.5

opencv-python-headless==4.12.0.88
ultralytics==8.3.220
pillow==11.3.0
scikit-image==0.25.2
imageio==2.37.0
imgaug==0.4.0

timm==1.0.21
transformers==4.57.1
huggingface-hub==0.36.0
tokenizers==0.22.1
safetensors==0.6.2
editdistance==0.8.1

numpy==2.2.6
lmdb==1.7.5
omegaconf==2.3.0
hydra-core==1.3.2
PyYAML==6.0.3

filterpy==1.4.5
scipy==1.16.2
matplotlib==3.10.7
shapely==2.1.2
httpx==0.28.1
requests==2.32.5
tqdm==4.67.1
regex==2025.10.23
packaging==25.0
nltk==3.9.2
fire==0.7.1
ray[default]==2.51.1